pyproject.toml
[project]
name = "nvfp4-quantization"
version = "0.1.0"
description = "NVFP4 quantization for Qwen3-VLTO-32B-Instruct on DGX Spark"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.5.0",
    "transformers>=4.46.0",
    "huggingface-hub>=0.26.0",
    "datasets>=3.1.0",
    "nvidia-modelopt>=0.18.0",
    "safetensors>=0.4.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
dev-dependencies = []
